requirements.txt
playwright==1.50.0
shapely==2.0.7
pyproj==3.7.1
folium==0.19.5
psycopg2-binary==2.9.10
pyyaml==6.0.2
